requirements.txt
google-cloud-bigquery>=3.25.0
jinja2>=3.1.4
anthropic>=0.34.0
requests>=2.32.0
pytest>=8.3.0
playwright>=1.48.0
